#117, #118, #119, #120 progress getting tests working
