+++ public/components/player/list/list_test.js
@@ -7,9 +7,8 @@ var fixture = require('can-fixture');
 
 F.attach(QUnit);
 
-
 var vm;
-QUnit.module("Player List Component", {
+QUnit.module("components/player/list/", {
 	beforeEach: function () {
 		localStorage.clear();
 		defineFixtures();
