+++ public/components/test.js
@@ -2,5 +2,6 @@ import './game/details/details_test';
 //import './navigation/navigation_test';
 //import './player/edit/edit_test';
 import './player/list/list_test';
-import './tournament/test';
+import './tournament/details/details_test';
+import './tournament/list/list_test';
 import './user/user_test';
