+++ public/components/tournament/details/details_test.js
@@ -4,7 +4,7 @@ import defineFixtures, { tournaments } from 'bitballs/models/fixtures/tournament
 
 var ViewModel = details.ViewModel;
 
-QUnit.module('Tournament Details ViewModel', {
+QUnit.module('components/tournament/details/', {
     beforeEach: function () {
         localStorage.clear();
         defineFixtures();
