+++ public/components/tournament/list/test.html
@@ -1,3 +1,3 @@
 <title>Tournament Tests</title>
-<script src="../../node_modules/steal/steal.js" main="bitballs/components/tournament/test"></script>
+<script src="../../../node_modules/steal/steal.js" main="bitballs/components/tournament/list/list_test"></script>
 <div id="qunit-fixture"></div
